Moved resolveSeparateDeps into dedicated trait
