+++ mill-mima/src/com/github/lolgab/mill/mima/ExtraCoursierSupport.scala
@@ -1,0 +1,43 @@
+package com.github.lolgab.mill.mima
+
+import mill.{Agg, PathRef, T}
+import mill.api.Result
+import mill.define.Task
+import mill.scalalib.{CoursierModule, Dep, Lib}
+
+trait ExtraCoursierSupport extends CoursierModule {
+
+  /** Resolves each dependency independently.
+    *
+    * @param deps
+    *   The dependencies to resolve.
+    * @param sources
+    *   If `true`, resolved the source jars instead of the binary jars.
+    * @return
+    *   Tuples containing each dependencies and it's resolved transitive
+    *   artifacts.
+    */
+  protected def resolveSeparateDeps(
+      deps: Task[Agg[Dep]],
+      sources: Boolean = false
+  ): Task[Agg[(Dep, Agg[PathRef])]] = T.task {
+    val pRepositories = repositoriesTask()
+    val pDepToDependency = resolveCoursierDependency().apply(_)
+    val pDeps = deps()
+    val pMapDeps = mapDependencies()
+    // API only available for mill 0.10 line
+    //    val pCustomizer = resolutionCustomizer()
+    pDeps.map { dep =>
+      val Result.Success(resolved) = Lib.resolveDependencies(
+        repositories = pRepositories,
+        depToDependency = pDepToDependency,
+        deps = Agg(dep),
+        sources = sources,
+        mapDependencies = Some(pMapDeps),
+        //        customizer = pCustomizer,
+        ctx = Some(implicitly[mill.api.Ctx.Log])
+      )
+      (dep, resolved)
+    }
+  }
+}
